# reports/openapi.yaml
openapi: 3.0.3
info:
  title: Kapda Co. API
  version: 1.0.0
  description: E-commerce platform API for designer marketplace
  contact:
    name: API Support
    email: api@kapdaco.com
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.kapdaco.com/api
    description: Production server
...
paths:
...
  /cart:
    get:
      tags: [Cart]
      summary: Get user cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Cart]
      summary: Add item to cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId:
                  type: string
                quantity:
                  type: integer
                  default: 1
                size:
                  type: string
                color:
                  type: string
      responses:
        '201':
          description: Item added to cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Cart:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              title:
                type: string
              price:
                type: number
              quantity:
                type: integer
              size:
                type: string
              color:
                type: string
        total:
          type: number
        updatedAt:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Access denied. No token provided.
...